+++ crimemap/docs/changes.md
@@ -10,3 +10,4 @@ Changes
 * 20-Mar-2014 --- Version 0.1.6 --- updated introduction.
 * 25-Mar-2014 --- Version 0.1.7 --- moved code to [https://github.com/woobe/rApps](https://github.com/woobe/rApps)
 * 04-Apr-2014 --- Version 0.1.8 --- added links to new personal website [http://www.jofaichow.co.uk](http://bit.ly/jofaichow)
+* 28-May-2014 --- Version 0.2.0 --- repalced "RJSONIO" with "jsonlite" for easier JSON reformatting. Fixed issues with API (data.police.uk changed the name of columns). Updated default demo to London Eye, 2013-01, 6 months, high res and BW.